+++ core/RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ## Version 3.3.1
 * Modified `core.playback.gapListObj.dump()` routine to return JSON element when given `filename='JSON'` as an argument.
+* Added log file path flag to `rsn_gaps.py`.
 
 ## Version 3.3.0
 * Added functionality to detect gaps in `rsn_gaps.py` which is built off the rest of the package.